backend/pyproject.toml
[project]
name = "mereb-gprc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "fastapi",
    "python-multipart>=0.0.20",
    "uvicorn",
    "black>=25.9.0",
    "python-dotenv",
    "celery",
    "redis",
    "ipdb>=0.13.13",
    "isort>=7.0.0",
    "pytest",
    "ruff>=0.14.3",
]
